+++ grid2op/Action/playableAction.py
@@ -109,7 +109,7 @@ class PlayableAction(BaseAction):
             self._switch_line_status,
             self._set_topo_vect,
             self._change_bus_vect,
-            self._redispatch,
+            self._private_redispatch,
             self._storage_power,
             {},
         )