+++ grid2op/tests/test_n_busbar_per_sub.py
@@ -8,6 +8,7 @@
 
 import warnings
 import unittest
+import numpy as np
 from grid2op.tests.helper_path_test import *
 
 import grid2op
@@ -18,7 +19,7 @@ from grid2op.Backend import PandaPowerBackend
 from grid2op.Space import DEFAULT_N_BUSBAR_PER_SUB
 from grid2op.Action import ActionSpace, BaseAction, CompleteAction
 from grid2op.Observation import BaseObservation
-from grid2op.Exceptions import Grid2OpException, EnvError, IllegalAction
+from grid2op.Exceptions import Grid2OpException, EnvError, AmbiguousAction
 from grid2op.gym_compat import GymEnv, DiscreteActSpace, BoxGymActSpace, BoxGymObsSpace, MultiDiscreteActSpace
 import pdb
 
@@ -71,14 +72,17 @@ class TestRightNumberNbBus(unittest.TestCase):
         
     def test_fail_if_not_int(self):
         with self.assertRaises(Grid2OpException):
-            env = grid2op.make("l2rpn_case14_sandbox", backend=_AuxFakeBackendSupport(), test=True, n_busbar="froiy", _add_to_name=type(self).__name__+"_wrong_str")
+            _ = grid2op.make("l2rpn_case14_sandbox", backend=_AuxFakeBackendSupport(), test=True, n_busbar="froiy", _add_to_name=type(self).__name__+"_wrong_str")
         with self.assertRaises(Grid2OpException):
-            env = grid2op.make("l2rpn_case14_sandbox", backend=_AuxFakeBackendSupport(), test=True, n_busbar=3.5, _add_to_name=type(self).__name__+"_wrong_float")
+            _ = grid2op.make("l2rpn_case14_sandbox", backend=_AuxFakeBackendSupport(), test=True, n_busbar=3.5, _add_to_name=type(self).__name__+"_wrong_float")
             
     def test_regular_env(self):
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            env = grid2op.make("l2rpn_case14_sandbox", backend=_AuxFakeBackendSupport(), test=True, _add_to_name=type(self).__name__+"_2")
+            env = grid2op.make("l2rpn_case14_sandbox",
+                               backend=_AuxFakeBackendSupport(),
+                               test=True,
+                               _add_to_name=type(self).__name__+"_2")
         self._aux_fun_test(env, DEFAULT_N_BUSBAR_PER_SUB)
         
         with warnings.catch_warnings():
@@ -503,6 +507,7 @@ class TestAction_3busbars(unittest.TestCase):
         else:
             # el not in topo vect (eg shunt)
             assert "shunt" in act_as_dict
+            assert "shunt_bus" in act_as_dict["shunt"]
             tmp = act_as_dict["shunt"]["shunt_bus"]
             assert tmp[el_id] == bus_val
 
@@ -516,6 +521,7 @@ class TestAction_3busbars(unittest.TestCase):
         else:
             # shunts of other things not in the topo vect
             assert "shunt" in act_as_dict
+            assert "shunt_bus" in act_as_dict["shunt"]
             tmp = act_as_dict["shunt"]["shunt_bus"]
             assert tmp == [(nm_els[el_id], bus_val)]
     
@@ -538,7 +544,7 @@ class TestAction_3busbars(unittest.TestCase):
         for bus in range(type(self.env).n_busbar_per_sub):
             self._aux_test_set_bus_onebus("load_set_bus", 0, bus + 1, type(self.env).name_load, 'loads_id')
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act.load_set_bus = [(0, type(self.env).n_busbar_per_sub + 1)]
             
     def test_set_gen_bus(self):
@@ -546,7 +552,7 @@ class TestAction_3busbars(unittest.TestCase):
         for bus in range(type(self.env).n_busbar_per_sub):
             self._aux_test_set_bus_onebus("gen_set_bus", 0, bus + 1, type(self.env).name_gen, 'generators_id')
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act.gen_set_bus = [(0, type(self.env).n_busbar_per_sub + 1)]
     
     def test_set_storage_bus(self):
@@ -554,7 +560,7 @@ class TestAction_3busbars(unittest.TestCase):
         for bus in range(type(self.env).n_busbar_per_sub):
             self._aux_test_set_bus_onebus("storage_set_bus", 0, bus + 1, type(self.env).name_storage, 'storages_id')
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act.storage_set_bus = [(0, type(self.env).n_busbar_per_sub + 1)]
     
     def test_set_lineor_bus(self):
@@ -562,7 +568,7 @@ class TestAction_3busbars(unittest.TestCase):
         for bus in range(type(self.env).n_busbar_per_sub):
             self._aux_test_set_bus_onebus("line_or_set_bus", 0, bus + 1, type(self.env).name_line, 'lines_or_id')
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act.line_or_set_bus = [(0, type(self.env).n_busbar_per_sub + 1)]
             
     def test_set_lineex_bus(self):
@@ -570,7 +576,7 @@ class TestAction_3busbars(unittest.TestCase):
         for bus in range(type(self.env).n_busbar_per_sub):
             self._aux_test_set_bus_onebus("line_ex_set_bus", 0, bus + 1, type(self.env).name_line, 'lines_ex_id')
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act.line_ex_set_bus = [(0, type(self.env).n_busbar_per_sub + 1)]
     
     def _aux_test_set_bus_onebus_sub_setbus(self, nm_prop, sub_id, el_id_sub, bus_val, name_xxx, el_nms):
@@ -591,7 +597,7 @@ class TestAction_3busbars(unittest.TestCase):
         for bus in range(type(self.env).n_busbar_per_sub):
             self._aux_test_set_bus_onebus_sub_setbus("sub_set_bus", 1, 0, bus + 1, type(self.env).name_line, 'lines_ex_id')
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act.line_ex_set_bus = [(0, type(self.env).n_busbar_per_sub + 1)]
             
     def test_change_deactivated(self):
@@ -623,7 +629,7 @@ class TestAction_3busbars(unittest.TestCase):
             self._aux_test_action_shunt(act, el_id, bus_val + 1)
             
         act = self.env.action_space()
-        with self.assertRaises(IllegalAction):
+        with self.assertRaises(AmbiguousAction):
             act = self.env.action_space({"shunt": {"set_bus": [(el_id, type(self.env).n_busbar_per_sub + 1)]}})